rt.bpel.server, AeEngineLifeCycleWrapper: dump default config to File before loading
The original code actually thought that files inside JARs could be opened just
like that. That doesn't work, as a file in a JAR is not a file to the
underlying OS. What does work, however, is dumping it to a temporary file and
opening that.

The problem with this approach for running the main ActiveBPEL WAR from the
Jetty Maven plugin is that it seems to require putting the aeEngineConfig.xml
file in the o.a.rt.bpel.server project, and I don't like having duplicate
content in the main code. For the time being, it's best to explicitly tell
ActiveBPEL to load a configuration stored in the testing part of the WAR
project.