o.a.rt.jetty: create subdir distributions in WAR projects
This allows us to use a predefined assembly descriptor in the Jetty project, and
run the Jetty project right from Eclipse, for easier debugging. To avoid repeating
ourselves too much, the assembly descriptor is defined in its own module.

We had to update maven-assembly-plugin from 2.1 to 2.2-beta-5 since properties
were not correctly interpolated until 2.2-beta-2, according to this bug report:

http://jira.codehaus.org/browse/MASSEMBLY-189